Prefect flow for VIIRS smoke presence and intensity.
